[feat#19] 컴포넌트화 수정 및 로딩 버튼 추가
[feat#19] 컴포넌트화 수정 및 로딩 버튼 추가
