+++ src/components/LineChart.jsx
@@ -37,7 +37,7 @@ const options = {
 
     scales: {
       y: { // [y 축 관련 설정] 
-        // min: 0, // [y 축 데이터 설정 0 ~ 30 까지 제한]
+        // min: 0, 
         // max: 30,
         grid: { // [y 축 데이터 시트 배경 선색 표시]
           drawBorder: false,
@@ -47,8 +47,8 @@ const options = {
         },
         
         ticks: {
-          color: '#D3D3D3',// [y 축 폰트 색상 설정]
-          font: { // [y축 폰트 스타일 변경]
+          color: '#D3D3D3',
+          font: {
             color : '#D3D3D3'
           } 
         }
@@ -57,8 +57,8 @@ const options = {
         display: true,
         offset: true,// [x 축 관련 설정] 
         ticks: {
-          color: '#808080', // [x 축 폰트 색상 설정]
-          font: { // [x축 폰트 스타일 변경]
+          color: '#808080', 
+          font: { 
             color : '#D3D3D3'
           } 
         },
@@ -210,7 +210,7 @@ const LineChart =() => {
     },[])
    
     return(
-        <div style={{width:'83%', height:'70%',margin: '20px'}}>
+        <div style={{width:'89%', height:'70%',margin: '20px'}}>
             <Bar data={data} options={options}/>
          </div>)
 }
