+++ src/pages/challenge/challenge.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Challenge from "../../components/Challenge"
+
+const challenge = () => {
+  return (
+    <>
+    <Challenge></Challenge>
+    </>
+  )
+}
+
+export default challenge
