+++ src/pages/global/Sidebar.jsx
@@ -188,7 +188,6 @@ const Sidebar = () => {
               color={colors.grey[300]}
               sx={{ m: "15px 0 5px 20px" }}
             >
-            
             </Typography>
             <LItem
               title="로그아웃"
