+++ src/pages/income/index.jsx
@@ -2,7 +2,6 @@ import { Box, useTheme } from "@mui/material";
 import { DataGrid } from "@mui/x-data-grid";
 import { tokens } from "../../theme";
 import { useState, useEffect, useRef } from "react";
-import 'bootstrap/dist/css/bootstrap.min.css';
 
 import axios from "axios";
 import Header from "../../components/Header";
@@ -11,6 +10,8 @@ import Form from 'react-bootstrap/Form';
 import Modal from 'react-bootstrap/Modal';
 import Sidebar from '../global/Sidebar';
 import Topbar from '../global/Topbar';
+import IncomeAddModal from '../../components/IncomeAddModal'
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 const Income = () => {
   
@@ -22,13 +23,10 @@ const Income = () => {
   const theme = useTheme();
   const [row, setRow] = useState("");
   const colors = tokens(theme.palette.mode);
-  const [show, setShow] = useState(false);
   const [isSidebar, setIsSidebar] = useState(true);
   const [editShow, setEditShow] = useState(false);
 
   const [list,setlist] = useState([]);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
   const handleEditClose = () => setEditShow(false);
   const handleEditShow = () => setEditShow(true);
 
@@ -42,27 +40,8 @@ const Income = () => {
       cost: "", 
   })
   
-  // 수입 내역 POST
-  function submit(e){
-    let user_id = localStorage.getItem("user_id")
-    axios.post('http://127.0.0.1:8000/api/v1/income/new/',{
-      user : user_id,
-      when: data.when,
-      memo: data.memo,
-      cost: data.cost
-    })
-      .then(res => {
-        memoRef.current.value = "";
-        costRef.current.value = "";
-        whenRef.current.value = "";
-        setlist([...list, ...res.data])
-      }) 
-      .catch(function (error) {
-        console.log(error);
-      });
-  }
 
-  // 지출 내역 화면 출력
+  // 수입 내역 화면 출력
   function handle(e){
     const newdata = {...data}
     newdata[e.target.id] = e.target.value
@@ -70,7 +49,7 @@ const Income = () => {
     console.log(newdata)
   }
 
-  // 지출 내역 DELETE
+  // 수입 내역 DELETE
   const handleDelete = (id)=>{
     if(window.confirm("삭제를 원하시면 확인 버튼을 눌러주세요.")){
       axios.delete(`http://127.0.0.1:8000/api/v1/income/${id}`)
@@ -84,7 +63,7 @@ const Income = () => {
   }
 
 
-// 지출 내역 UPDATE
+ // 수입내역 UPDATE
   const handleEdit= (id) => {
     axios.put(`http://127.0.0.1:8000/api/v1/income/${id}`,{
       user : user_id,
@@ -142,7 +121,6 @@ const Income = () => {
             </>
         )}
     }
-
   ];
   
   // 수입 내역 GET
@@ -232,74 +210,12 @@ const Income = () => {
           }
        `}
       </style>
-      <Button variant="primary main" onClick={handleShow}>
-            내역 추가
-      </Button>
-      
-      <Modal show={show} onHide={handleClose}>
-        <Modal.Header closeButton>
-          <Modal.Title>수입 내역 추가</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <Form>
-            <Form.Group className="mb-3" >
-              <Form.Label>날짜</Form.Label >
-              <Form.Control
-                ref={whenRef}
-                type="date"
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="when"
-                value ={data.when}
-                method="post"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3" >
-            <Form.Label>메모</Form.Label>
-              <Form.Control
-                ref={memoRef}
-                type="text"
-                placeholder="메모 (최대 50자)"
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="memo"
-                value ={data.memo}
-                method="post"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3" >
-            <Form.Label>금액</Form.Label>
-              <Form.Control
-                ref={costRef}
-                type="text"
-                placeholder="금액을 입력하세요."
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="cost"
-                value ={data.cost}
-                method="post"
-              />
-            </Form.Group>
-          </Form>
-        </Modal.Body>
-        <Modal.Footer>
-          
-          <Button variant="secondary center" onClick={handleClose}>
-            닫기
-          </Button>
-          <Button type="submit" variant="primary second"
-           onClick={()=> {
-            handleClose();
-            submit();
-            window.location.reload()
-            }}> 확인 </Button>
-        </Modal.Footer>
 
-      </Modal>
+      <IncomeAddModal></IncomeAddModal>
 
-     <Modal show={editShow} onHide={handleClose}>
+     <Modal show={editShow} onHide={handleEditClose}>
         <Modal.Header closeButton>
-          <Modal.Title>지출 내역 수정</Modal.Title>
+          <Modal.Title>수입 내역 수정</Modal.Title>
         </Modal.Header>
         <Modal.Body>
           <Form>
