+++ src/pages/spending/index.jsx
@@ -11,12 +11,14 @@ import Form from 'react-bootstrap/Form';
 import Modal from 'react-bootstrap/Modal';
 import Sidebar from '../global/Sidebar';
 import Topbar from '../global/Topbar';
+import SpendingAddModal from '../../components/SpendingAddModal'
 
 const Spending = () => {
   
   let today = new Date();  
   let year = today.getFullYear(); // 년도
   let month = today.getMonth() + 1;  // 월
+  let formData = new FormData();
 
   let user_id = localStorage.getItem("user_id")
   const theme = useTheme();
@@ -28,12 +30,10 @@ const Spending = () => {
   const [isSidebar, setIsSidebar] = useState(true);
   const [list,setlist] = useState([]);
   const [file, setFile] = useState(false);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
 
+  const handleShow = () => setShow(true);
   const handleEditClose = () => setEditShow(false);
   const handleEditShow = () => setEditShow(true);
-
 
   const memoRef = useRef()
   const costRef = useRef()
@@ -46,28 +46,6 @@ const Spending = () => {
       purpose: "",
       cost: "", //초기값
   })
-
-  
-  // 지출 내역 POST
-  function submit(e){
-    axios.post('http://127.0.0.1:8000/api/v1/spending/new/',{
-      user : user_id,
-      when: data.when,
-      memo: data.memo,
-      purpose: data.purpose,
-      cost: data.cost
-    })
-      .then(res => {
-        memoRef.current.value = "";
-        costRef.current.value = "";
-        whenRef.current.value = "";
-        purposeRef.current.value = "";
-        setlist([...list, ...res.data])
-      }) 
-      .catch(function (error) {
-        console.log(error);
-      });
-  }
 
   // 지출 내역 화면 출력
   function handle(e){
@@ -82,9 +60,8 @@ const Spending = () => {
       setFile(event.target.files[0]);
     };
 
-// ocr 업로드 버튼 클릭시
+  // ocr 업로드 버튼 클릭시
   const upload = () => {
-    let formData = new FormData();
     formData.append("files", file);
     axios({
       method: "post",
@@ -115,12 +92,10 @@ const Spending = () => {
     }
   }
 
-
-// 지출 내역 UPDATE
+  // 지출 내역 UPDATE
   const handleEdit= (id) => {
     axios.put(`http://127.0.0.1:8000/api/v1/spending/${id}`,{
-      user : user_id, //로그인한 user_id 보내줘야함. 세션에서 user_id를 꺼내와야함
-      //로컬 스토리지에 user_id를 저장해서, user_id를 비교해서 
+      user : user_id,
       when: data.when,
       memo: data.memo,
       purpose: data.purpose,
@@ -137,9 +112,17 @@ const Spending = () => {
         console.log(error);
       });
     }
+   // 지출 내역 GET
+    useEffect(() => {
+      const user_id = localStorage.getItem("user_id")
+      axios.get(`http://127.0.0.1:8000/api/v1/spending/spending-list/${user_id}`)
+      .then(res => {
+        setlist(res.data.spending_list);
+    })
+  },[])
 
 
-  // DataGrid 열 (날짜, 용도, 메모, 금액)
+  // DataGrid Columns (날짜, 용도, 메모, 금액)
   const columns = [
     {
       field: "when",
@@ -187,17 +170,7 @@ const Spending = () => {
             </>
         )}
     }
-
   ];
-  
-  // 지출 내역 GET
-  useEffect(() => {
-    const user_id = localStorage.getItem("user_id")
-    axios.get(`http://127.0.0.1:8000/api/v1/spending/spending-list/${user_id}`)
-    .then(res => {
-      setlist(res.data.spending_list);
-    })
-  },[])
 
   return (
     <div className="app" >
@@ -208,36 +181,6 @@ const Spending = () => {
     <div className="size">
     <Box m="20px">
       <Header title="지출내역" subtitle={(year + '년 '+ month + '월')}/>
-      <Box
-        m="40px 0 0 0"
-        height="75vh"
-        sx={{
-          "& .MuiDataGrid-root": {
-            border: "none",
-          },
-          "& .MuiDataGrid-cell": {
-            borderBottom: "none",
-          },
-          "& .name-column--cell": {
-            color: colors.greenAccent[300],
-          },
-          "& .MuiDataGrid-columnHeaders": {
-            backgroundColor: colors.blueAccent[700],
-            borderBottom: "none",
-          },
-          "& .MuiDataGrid-virtualScroller": {
-            backgroundColor: colors.primary[400],
-          },
-          "& .MuiDataGrid-footerContainer": {
-            borderTop: "none",
-            backgroundColor: colors.blueAccent[700],
-          },
-          "& .MuiCheckbox-root": {
-            color: `${colors.greenAccent[200]} !important`,
-          },
-        }}
-      >
-      
       <style type="text/css">
         {`
           .btn.ListEdit {
@@ -278,102 +221,42 @@ const Spending = () => {
           }
        `}
       </style>
-      <Button variant="primary main" onClick={handleShow}>
-            내역 추가
-      </Button>
-      
-      <Modal show={show} onHide={handleClose}> {/*post*/}
-        <Modal.Header closeButton>
-          <Modal.Title>지출 내역 추가</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <Form>
-            <Form.Group className="mb-3" >
-              <Form.Label>날짜</Form.Label >
-              <Form.Control
-                ref={whenRef}
-                type="date"
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="when"
-                value ={data.when}
-                method="post"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3" >
-            <Form.Label>메모</Form.Label>
-              <Form.Control
-                ref={memoRef}
-                type="text"
-                placeholder="메모 (최대 50자)"
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="memo"
-                value ={data.memo}
-                method="post"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3">
-            <Form.Label>용도</Form.Label>
-              <select class="form-select" 
-              ref={purposeRef}
-              onChange={(e) => handle(e)}
-              id ="purpose"
-              value ={data.purpose}
-              method="post">
-                <option selected>용도를 선택하세요.</option>
-                <option value="식사">식사</option>
-                <option value="술/유흥">술/유흥</option>
-                <option value="뷰티/미용">뷰티/미용</option>
-                <option value="교통/차량">교통/차량</option>
-                <option value="주거/통신">주거/통신</option>
-              </select>
-            </Form.Group>
-            <Form.Group className="mb-3" >
-            <Form.Label>금액</Form.Label>
-              <Form.Control
-                ref={costRef}
-                type="number"
-                placeholder="금액을 입력하세요."
-                autoFocus
-                onChange={(e) => handle(e)}
-                id ="cost"
-                value ={data.cost}
-                method="post"
-              />
-            </Form.Group>
-            <Form.Group className="mb-3">
-            <Form.Label>영수증 OCR</Form.Label>
-              <Form.Control
-                type="file"
-                autoFocus
-                accept="image/*" 
-                onChange={handleInputChange} 
-                method="post"
-              />
-               <Button variant="primary center"
-                onClick={() => {
-                handleShow();
-                upload();
-               }}> 업로드 </Button> <br/>
-               <Form.Label className="m">❗영수증 이미지를 업로드 하면, 자동으로 가계부를 작성해드립니다❗</Form.Label>
-            </Form.Group>
-          </Form>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary center" onClick={handleClose}>
-            닫기
-          </Button>
-          <Button type="submit" variant="primary second"
-           onClick={()=> {
-            handleClose();
-            submit();
-            window.location.reload()
-            }}> 확인 </Button>
-        </Modal.Footer>
-      </Modal>
 
-      <Modal show={editShow} onHide={handleClose}>
+      <Box
+        m="40px 0 0 0"
+        height="75vh"
+        sx={{
+          "& .MuiDataGrid-root": {
+            border: "none",
+          },
+          "& .MuiDataGrid-cell": {
+            borderBottom: "none",
+          },
+          "& .name-column--cell": {
+            color: colors.greenAccent[300],
+          },
+          "& .MuiDataGrid-columnHeaders": {
+            backgroundColor: colors.blueAccent[700],
+            borderBottom: "none",
+          },
+          "& .MuiDataGrid-virtualScroller": {
+            backgroundColor: colors.primary[400],
+          },
+          "& .MuiDataGrid-footerContainer": {
+            borderTop: "none",
+            backgroundColor: colors.blueAccent[700],
+          },
+          "& .MuiCheckbox-root": {
+            color: `${colors.greenAccent[200]} !important`,
+          },
+        }}
+      >
+
+      {/* 지출내역 추가 */}
+      <SpendingAddModal></SpendingAddModal>
+      
+      {/* 지출내역 수정 */}
+      <Modal show={editShow} onHide={handleEditClose}>
         <Modal.Header closeButton>
           <Modal.Title>지출 내역 수정</Modal.Title>
         </Modal.Header>
@@ -463,19 +346,14 @@ const Spending = () => {
             handleEdit(row);         
             window.location.reload()
             }}> 확인 </Button>
-            
-          
         </Modal.Footer>
       </Modal>
+
         <DataGrid 
         onSelectionModelChange={datas => {console.log(datas.toString())}}
         checkboxSelection rows={list}
         columns={columns}
         getRowId={list => list.id} />
-        {/* onChange={(e) => {setlist(e.target.list)}}
-        {list.map(spending => (
-          <SpendingItem spending={spending}/>
-        ))} */}
       </Box>
     </Box>
     </div>
@@ -484,9 +362,3 @@ const Spending = () => {
 };
 
 export default Spending;
-// function SpendingItem({spending}) {
-//     return ( <div>
-//       <span>{spending.id}</span>
-//       <button type='button' onClick={() => alert(`Spending id clicked: ${spending.id}`)}>CLICK</button>
-//     </div>)
-//   }
